refactor(appserver): move rpc+lifecycle into core
